requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.0
jinja2==3.1.4
python-multipart==0.0.9
pystray==0.19.5
Pillow>=9.0.0,<11.0.0
requests>=2.28.0,<3.0.0